Lock version numbers
We don't want updates to the libraries to break our deployments. I
locked these version numbers by first making sure that none of the
versions contain a caret(^) symbol in front of them. The caret gets the
latest version that is under the same major version according to
semantic versioning. For libraries that had a caret, I applied the caret
logic to them manually, then removed the caret to lock the version. An
exception to this is react, react-dom, and react-addons-test-utils
libraries. The latest version under the same major version was the
culprit that broke our deployments.