Change the way we get Localize to use url
Using the version number requires us to change the npm registry to point
to a different url before we are able to npm install the library. This
might introduce bugs since other libraries that we use are from the
default npm registry url. We use a direct url that points to the path in
artifactory for Localize to get the library.